semantic_server/requirements.txt
fastapi
uvicorn
open_clip_torch
torch
pillow
sentence-transformers
pymupdf
python-docx
python-pptx
numpy